+++ gherkin/dart/lib/parser.dart
@@ -1,0 +1,8 @@
+library gherkin.parser;
+
+export 'src/parser/IBuilder.dart';
+export 'src/parser/ITokenMatcher.dart';
+export 'src/parser/ITokenScanner.dart';
+export 'src/parser/Parser.dart';
+export 'src/parser/ParserContext.dart';
+export 'src/parser/RuleType.dart';